fix python solutions
