--- compiled_starters/sqlite-starter-rust/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod header;
-pub mod record;
-pub mod schema;
-pub mod varint;
