--- solutions/rust/01-init/code/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod header;
-pub mod record;
-pub mod schema;
-pub mod varint;
