+++ solutions/rust/01-init/code/src/main.rs
@@ -6,6 +6,17 @@ use std::convert::TryInto;
 use std::fs::File;
 use std::io::prelude::*;
 
+/// parses variable size unsigned integer encoded as bytes to number.
+fn parse_number(bytes: &[u8]) -> usize {
+    let mut result: usize = 0;
+    let num_bytes = bytes.len();
+    for i in 0..num_bytes {
+        let shift = (num_bytes - i - 1) * 8;
+        result += (bytes[i] as usize) << shift;
+    }
+    result
+}
+
 fn main() -> Result<()> {
     // Parse arguments
     let args = std::env::args().collect::<Vec<_>>();
@@ -15,39 +26,18 @@ fn main() -> Result<()> {
         _ => {}
     }
 
-    // Read database file into database
-    let mut file = File::open(&args[1])?;
-    let mut database = Vec::new();
-    file.read_to_end(&mut database)?;
-
     // Parse command and act accordingly
     let command = &args[2];
     match command.as_str() {
         ".dbinfo" => {
-            // Parse page header from database
-            let page_header = PageHeader::parse(&database[100..108])?;
+            let mut file = File::open(&args[1])?;
+            let mut header = [0; 100];
+            file.read_exact(&mut header)?;
 
-            // Obtain all cell pointers
-            let cell_pointers = database[108..]
-                .chunks_exact(2)
-                .take(page_header.number_of_cells.into())
-                .map(|bytes| u16::from_be_bytes(bytes.try_into().unwrap()))
-                .collect::<Vec<_>>();
+            // The page size is stored at the 16th byte offset, using 2 bytes in little-endian order
+            let page_size = u16::from_le_bytes([header[16], header[17]]);
 
-            // Obtain all records from column 5
-            #[allow(unused_variables)]
-            let schemas = cell_pointers
-                .into_iter()
-                .map(|cell_pointer| {
-                    let stream = &database[cell_pointer as usize..];
-                    let (_, offset) = parse_varint(stream);
-                    let (_rowid, read_bytes) = parse_varint(&stream[offset..]);
-                    parse_record(&stream[offset + read_bytes..], 5)
-                        .map(|record| Schema::parse(record).expect("Invalid record"))
-                })
-                .collect::<Result<Vec<_>>>()?;
-
-            println!("number of tables: {}", schemas.len());
+            println!("database page size: {}", page_size);
         }
         _ => bail!("Missing or invalid command passed: {}", command),
     }
